Only allow Openable content in Storage Access Framework
Add the CATEGORY_OPENABLE to the Intent to launch the Storage Access Framework (SAF) when sending a file. Setting this category ensures that the OpenableColumns data contract is met by the opened data, which we rely upon during data upload.

In practice, some content, such as Contacts, will no longer display in the SAF UI.

Fixes #935