Remove the request for READ_EXTERNAL_STORAGE permission
We use the Storage Access Framework (SAF) to enable users to send files. When using the SAF, READ_EXTERNAL_STORAGE permission is not needed to read from internal storage, SD Cards, or anything else. The Android System is the one accessing the storage, so applications do not need direct access. This is in fact one of the major selling points of the SAF.

Note that no Google samples or apps by Google Developer Advocates that highlight the SAF request this permission.

Removing this permission is a good thing. Firstly, there are many SAF providers such as Google Drive, Photos, Microsoft OneDrive, etc, that do not use external storage. Secondly, users are not prompted for a needless permission. Thirdly, it reduces the complexity of the ChatRoomFragment.