Fix file uploading when no filters are applied
When no file filters are specified in the server whitelist, the Android system is incorrectly given a mime type filter list including a single empty ("") mime type. Naturally, nothing matches this, and no documents can be selected for upload.

To remedy this, the application is changed to return a null filter array if filters are not to be applied.