Change ImagePipelineConfig initialization in Dagger graph
Rather than creating a new, unsized HashSet, use idiomatic Kotlin to create the set of RequestListeners for the ImagePipelineConfig. Under the hood a SingletonSet will be used which is optimized for the current use case, but if additional Listeners are added, the collections will automatically be sized appropriately. This leads to better memory utilization and reduced CPU usage.