Remove the FrescoAuthInterceptor
The FrescoAuthInterceptor has not been used since it was introduced, suggesting that it is not required for proper function. By not needing the FrescoAuthInterceptor, it is not necessary to copy or store the OkHttpClient.

Removing the FrescoAuthInterceptor produces a smaller APK due to there being less code. It reduces app cold start times by reducing the CPU time spent initializing an unused class and copying the OkHttpClient. It reduces RAM consumption by not keeping a deep copy of the OkHttpClient around for the lifetime of the application. Finally, it reduces compile time as less time is spent performing Dagger graph analysis.