Fixed two issues:
1. recycler_view was adding its adapter in showMessages this was causing the error: "E/RecyclerView: No adapter attached; skipping layout"
This seems to be the cause of a missed update step and ui delay.  Code to add adapter was added to setupRecyclerView, which is called in the onViewCreated function.

2. loadActiveMembers function was being called from ChatRoomFragment object and the network / data access was being done in the UI thread.  Moved the function call from the end of the function loadMessages to just after the end of said function.
Changed the function loadActiveMembers to do most of its work in an I.O. thread and just the view updating in a UI thread.